feat: add force flag

+++ src/_utils/writeFile.ts
@@ -1,7 +1,15 @@
 import fs from "fs";
 
-export default function writeFile(path: string, data: string): Promise<void> {
-  return new Promise((resolve, reject) =>
+import exists from "./exists";
+
+export default async function writeFile(
+  path: string,
+  data: string,
+  force: boolean,
+): Promise<void> {
+  if (!force) if (await exists(path)) return;
+
+  return await new Promise((resolve, reject) =>
     fs.writeFile(path, data, { encoding: "utf-8" }, (error) =>
       error === null ? resolve() : reject(error),
     ),
